[BUG] Parse beers correctly
The way that this tool was parsing beers was broken, as the beer board changed location and also drastically changed in terms of mark-up.

I've moved the web-scraping and parsing logic out into a separate library (to make it more easily testable), see: https://github.com/TomSeldon/tiny-rebel-web-scraper

I've modified this tool to just use that web scraper, and in turn made it a lot simpler.

**Note:** I've removed the beer ratings (fetched via untapped) as I didn't really understand what it was doing, and frankly I want to spend as little time  working with Coffeescript as possible.

The format of the output has changed, though I'm really not precious about it. The goal here was to make it output something correct again, someone else is welcome to change exactly *what* is output.

I've also updated the tests to make use of a hubot testing library, making it easier to assert on the messages output. The test(s) really aren't great, but is at least some validation that the tool works. The tests for the web scraping / parsing are a lot more complete.

As the web scraper library uses native promises, this may impact how this Hubot script is run. It will require a version of Node that supports native promises. See: http://node.green/#Promise

[CONFIG] Change what versions of Node are supported